+++ R/scan_project_files.R
@@ -1,19 +1,31 @@
 #' Scan R files for package dependencies
 #'
-#' This function scans all files containing R code, including scripts, Sweave documents and Rmarkdown-based files, for references to packages. It parses the code and looks for calls to `library` and `require`, as well as the namespacing operators `::` and `:::`.
+#' This function scans the R files in your project, including scripts, Sweave documents and Rmarkdown-based files, for references to packages.
 #'
 #' @param project_dir A project path.  This is the path to the root of the project that references the packages to be installed from the MRAN snapshot for the date specified for `snapshotDate`. Defaults to the current working directory.
 #'
 #' @param scan_r_only If `TRUE`, limits the scanning of project files to R scripts only (those with the extension `.R`).
 #'
-#' @param scan_rnw_with_knitr If `TRUE`, scans Sweave files (those with extension `.Rnw`) with [`knitr::knitr`], otherwise with [`utils::Stangle`]. Ignored if `scan_r_only=TRUE`.
+#' @param scan_rnw_with_knitr If `TRUE`, scans Sweave files with [`knitr::knitr`], otherwise with [`utils::Stangle`]. Ignored if `scan_r_only=TRUE`.
 #'
 #' @param scan_rprofile if `TRUE`, includes the `~/.Rprofile` startup file in the scan. See [`Startup`].
 #'
+#' @details
+#' `scan_project_files` recursively builds a list of all the R files in your project. This includes regular R scripts, as well as Sweave files (those with extension `.Rnw`) and Rmarkdown-based files (those with extension `.Rmd`, `.Rpres` or `Rhtml`). It then parses the code in each file and looks for calls to `library` and `require`, as well as the namespacing operators `::` and `:::`.
+#'
+#' @section Manifest:
+#'
+#' As an **experimental feature**, you can specify additional packages to include or exclude via an optional `checkpoint.yml` manifest file located in your project directory. This should be a valid YAML file with 2 components:
+#' - `refs`: An array of package references, that can be parsed by [`pkgdepends::new_pkg_installation_proposal`]. See [`pkgdepends::pkg_refs`] for a description of the reference syntax.
+#' - `exclude`: An array of package names (without decorations) that should be excluded from the final list of dependencies.
+#'
+#' A manifest file allows you to include packages from BioConductor or GitHub in the checkpoint. You should use this feature with caution, as checkpoint does not check the versions of these packages. While pkgdepends allows you to specify a package version as part of a reference, it currently ignores the version during the download and installation process.
+#'
+#' A use case for exclusions might be if your workflow requires packages that are not available from CRAN or other public repositories. For example, Microsoft Machine Learning Server (MMLS) comes with a number of proprietary packages for big data and in-database analytics. You can exclude these packages from checkpointing by listing them in the `exclude` list in the manifest. In this case, you must ensure that your packages are still visible to the checkpointed session, generally by modifying the value of `.libPaths()` manually. (For MMLS, this is easy as the proprietary packages live in the base library directory, which is always part of `.libPaths()`.)
 #' @return
 #' A list with 2 components: `pkgs`, a vector of package names, and `errors`, a vector of files that could not be scanned. The package listing includes third-party packages, as well as those that are distributed with R and have "Recommended" priority. Base-priority packages (utils, graphics, methods and so forth) are not included.
 #'
-#' In addition, if any Rmarkdown-based files are found (those with extension `.Rmd`, `.Rpres` or `Rhtml`), the package listing will include rmarkdown. This allows you to continue rendering them in a checkpointed session.
+#' In addition, if any Rmarkdown files are found, the package listing will include rmarkdown. This allows you to continue rendering them in a checkpointed session.
 #' @examples
 #'
 #' scan_project_files()
@@ -54,6 +66,7 @@ scan_project_files <- function(project_dir=".", scan_r_only=FALSE, scan_rnw_with
     if(length(errors) > 0)
         warning("Following files could not be scanned:\n", paste(errors, collapse="\n"), call.=FALSE)
 
+    mft <- read_manifest(project_dir)
     exclude <- c(
         # this package
         "checkpoint",
@@ -65,7 +78,8 @@ scan_project_files <- function(project_dir=".", scan_r_only=FALSE, scan_rnw_with
           "tools", "utils")
     )
     any_rmd <- any(grepl("\\.(rmd|rpres|rhtml)$", r_files, ignore.case=TRUE))
-    pkgs <- setdiff(unique(c(pkgs, if(any_rmd) "rmarkdown")), exclude)
+
+    pkgs <- setdiff(unique(c(pkgs, mft$refs, if(any_rmd) "rmarkdown")), c(mft$exclude, exclude))
     list(pkgs=pkgs, errors=errors)
 }
 
@@ -121,3 +135,16 @@ find_dependencies <- function(e)
         else return(character(0))
     }
 }
+
+read_manifest <- function(project_dir)
+{
+    mft_file <- file.path(project_dir, "checkpoint.yml")
+    if(!file.exists(mft_file))
+        mft_file <- file.path(project_dir, "checkpoint.yaml")
+    if(!file.exists(mft_file))
+        return(NULL)
+
+    mft <- yaml::read_yaml(mft_file)
+    mft[c("refs", "exclude")]
+}
+